chore: CE changes for module input autocomplete (#28221)
