+++ app/client/src/ee/utils/autocomplete/entityDefGeneratorMap.ts
@@ -1,0 +1,1 @@
+export * from "ce/utils/autocomplete/entityDefGeneratorMap";
